+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0 (UNRELEASED)
 
+* Added `--upscale` option to upscale the input image by an integer factor.
 * Fixed `snake`-mode bug where the last pixel (or group of pixel) would be missing in some cases.
 
 ## 0.2.0 (2022-04-28)